+++ src/ybot_manager.erl
@@ -74,6 +74,10 @@ handle_call({get_plugin, PluginName}, _From, State) ->
 handle_call(get_transports, _From, State) ->
     {reply, State#state.transports, State};
 
+%% @doc get channels info
+handle_call(get_channels, _From, State) ->
+    {reply, State#state.channels, State};
+
 %% @doc get all runned transports pid
 handle_call(get_runnned_transports, _From, State) ->
     % Return all runned transports